// temper/shared-addon-libraries-lib-addon-menu/src/types.ts
export type Valued<T> = T | ((this: void) => T)

export interface LamWidgetData {
  type: string
  name?: Valued<string | number>
  reference?: string
  width?: "full" | "half"
  tooltip?: Valued<string | number>
  tooltipText?: string | number
  warning?: Valued<string | number>
  requiresReload?: boolean
  disabled?: Valued<boolean>
  helpUrl?: Valued<string>
  default?: unknown
  resetFunc?: (this: void, control: LamControl) => void
  getFunc?: (this: void) => unknown
  setFunc?: (this: void, ...args: unknown[]) => void
  registerForRefresh?: boolean
  registerForDefaults?: boolean
  controls?: LamWidgetData[]
}

interface LamValueWidgetData extends LamWidgetData {
  getFunc: (this: void) => unknown
  setFunc: (this: void, ...args: unknown[]) => void
}

export interface PanelData extends LamWidgetData {
  type: "panel"
  name: Valued<string | number>
  displayName?: Valued<string | number>
  author?: Valued<string | number>
  version?: Valued<string | number>
  website?: Valued<string>
  feedback?: Valued<string>
  translation?: Valued<string>
  donation?: Valued<string>
  keywords?: string
  slashCommand?: string
  registerForRefresh?: boolean
  registerForDefaults?: boolean
}
// ...
export interface DropdownData extends LamValueWidgetData {
  type: "dropdown"
  name: Valued<string | number>
  choices: string[]
  choicesValues?: unknown[]
  choicesTooltips?: Valued<string | number>[]
  sort?: string
  scrollable?: boolean | number
  multiSelect?: Valued<boolean>
  multiSelectTextFormatter?: Valued<string | number>
  multiSelectNoSelectionText?: Valued<string | number>
  multiSelectMaxSelections?: Valued<number>
}
// ...
export type UpdateValueFn = (this: LamControl, forceDefault?: boolean, ...args: unknown[]) => void
export type UpdateFn = (this: LamControl) => void

export interface IconControl extends TextureControl {
  texture?: string
  tooltip?: string
  color?: ZoColorDef
  size?: number
  OnSelect?: (this: void, icon: Control, texture: string) => void
  SetTexture: (this: IconControl, texture: unknown) => void
}

export interface LamControl extends Control {
  data: LamWidgetData
  panel?: LamControl
  isHalfWidth?: boolean
  container?: Control
  label?: Control
  labelContainer?: Control
  faqControl?: Control
  scroll?: Control
  lineControl?: Control
  startValue?: unknown[]
  value?: unknown
  isDisabled?: boolean
  disabled?: boolean
  disabledLabel?: boolean
  warning?: Control
  info?: Control
  website?: Control
  feedback?: Control
  translation?: Control
  donation?: Control
  controlsToRefresh?: LamControl[]
  icon?: IconControl
  bg?: Control
  arrow?: Control
  animation?: TimelineAnimation
  open?: boolean
  btmToggle?: Control
  button?: Control
  checkbox?: Control
  checkedText?: string
  uncheckedText?: string
  color?: Control
  thumb?: Control
  desc?: Control
  title?: Control
  header?: Control
  divider?: Control
  texture?: Control
  editbox?: Control
  slider?: Control
  minText?: Control
  maxText?: Control
  slidervalueBG?: Control
  slidervalue?: Control
  choices?: Record<string | number, unknown>
  combobox?: Control
  dropdown?: ZoComboBox
  isMultiSelectionEnabled?: boolean
  m_sortType?: unknown
  m_sortOrder?: unknown
  comboboxCount?: number
  dropdownButton?: Control
  UpdateValue?: UpdateValueFn
  UpdateDisabled?: UpdateFn
  UpdateWarning?: UpdateFn
  RefreshPanel?: (this: LamControl) => void
  ForceDefaults?: (this: LamControl) => void
  UpdateChoices?: (this: LamControl, ...args: unknown[]) => void
  SetColor: (this: Control, ...args: unknown[]) => void
  SetIconSize?: (this: LamControl, size: number) => void
  SetDropdownHeight?: (
    this: LamControl,
    dropdown: ZoComboBox,
    dropdownData: DropdownData
  ) => LuaMultiReturn<[number, number, number]>
  AdjustDimensions?: (this: LamControl) => void
}

export interface IconPickerMenu {
  control: Control
  parent?: Control
  color: ZoColorDef
  refCount?: number
  customOnMouseEnter?: (this: void, icon: Control) => void
  customOnMouseExit?: (this: void, icon: Control) => void
  Clear: (this: IconPickerMenu) => void
  AddIcon: (
    this: IconPickerMenu,
    texturePath: string,
    callback: (this: void, icon: Control, texture: string) => void,
    tooltip?: string
  ) => void
  Show: (this: void, parent: Control) => boolean
  SetColor: (this: IconPickerMenu, color: ZoColorDef) => void
  SetMaxColumns: (this: void, value?: number) => void
  SetIconSize: (this: void, value?: number) => void
  SetVisibleRows: (this: void, value?: number) => void
  SetMouseHandlers: (
    this: void,
    onEnter?: (this: void, icon: Control) => void,
    onExit?: (this: void, icon: Control) => void
  ) => void
  UpdateDimensions: (this: void) => void
  UpdateAnchors: (this: void) => void
}

export interface TooltipData {
  tooltipText?: string | number
  helpUrl?: string
}

export interface TooltipHostControl extends Control {
  data?: TooltipData
}

export interface FaqTextureControl extends TextureControl {
  data?: TooltipData
}

export interface AddonListData {
  panel: LamControl
  name: string
  filterText: string
  sortIndex?: number
}

export interface LamUtil {
  L: import("./ui-strings").LamStrings
  GetTooltipText: (this: void, value: Valued<string | number>) => string | number
  GetStringFromValue: (this: void, value: Valued<string | number>) => string | number
  GetDefaultValue: <T>(this: void, value: Valued<T>) => T
  GetColorForState: (this: void, disabled: boolean) => ZoColorDef
  CreateBaseControl: (
    this: void,
    parent: LamControl,
    controlData: LamWidgetData,
    controlName?: string
  ) => LamControl
  CreateLabelAndContainerControl: (
    this: void,
    parent: LamControl,
    controlData: LamWidgetData,
    controlName?: string
  ) => LamControl
  SetUpTooltip: (
    this: void,
    control: Control,
    data: LamWidgetData,
    tooltipData?: TooltipData
  ) => void
  RequestRefreshIfNeeded: (this: void, control: LamControl) => void
  RegisterForRefreshIfNeeded: (this: void, control: LamControl) => void
  RegisterForReloadIfNeeded: (this: void, control: LamControl) => void
  GetTopPanel: (this: void, panel: LamControl) => LamControl
  ShowConfirmationDialog: (
    this: void,
    title: string,
    body: string,
    callback: (this: void, ...args: unknown[]) => void
  ) => void
  UpdateWarning: (this: void, control: LamControl) => void
  CreateFAQTexture: (this: void, control: LamControl) => Control | undefined
  GetIconPickerMenu?: (this: void) => IconPickerMenu
}

export interface Lam {
  widgets: Record<string, number>
  util: LamUtil
  controlsForReload: LamControl[]
  requiresReload?: boolean
  applyButton?: Control
  defaultButton?: Control
  addonList?: Control
  currentAddonPanel?: LamControl
  pendingAddonPanel?: LamControl
  currentPanelOpened?: boolean
  panelId?: number
  keybindsInitialized?: boolean
  RegisterWidget: (this: Lam, widgetType: string, widgetVersion: number) => boolean
  RegisterAddonPanel: (this: Lam, addonID: string, panelData: PanelData) => LamControl | undefined
  RegisterOptionControls: (this: Lam, addonID: string, optionsTable: LamWidgetData[]) => void
  OpenToPanel: (this: Lam, panel: LamControl) => void
  GetAddonPanelContainer: (this: Lam) => Control
  GetAddonSettingsFragment: (this: Lam) => ZoFadeSceneFragment
}

export type LamFactory = (
  this: void,
  parent: LamControl,
  data: LamWidgetData,
  controlName?: string
) => LamControl

export interface Lamcc {
  scrollCount: number
  comboboxCount?: number
  [widgetType: string]: LamFactory | number | undefined
}